+++ client/src/features/home/components/index.js
@@ -1,0 +1,3 @@
+export * from "./ChartAreaInteractive.jsx"
+export * from "./DataTable.jsx"
+export * from "./SectionCards.jsx"